Move logic for creating "mockAnalysisStore"s into the "core"
